Pull at 20180416

+++ dnSpy/Roslyn/dnSpy.Roslyn/Debugger/FilterExpressionEvaluator/EvalDelegateCreator.cs
@@ -26,7 +26,7 @@ using SR = System.Reflection;
 using SRE = System.Reflection.Emit;
 
 namespace dnSpy.Roslyn.Debugger.FilterExpressionEvaluator {
-	delegate bool EvalDelegate(string machineName, ulong processId, string processName, ulong threadId, string threadName);
+	delegate bool EvalDelegate(string machineName, int processId, string processName, ulong threadId, string threadName);
 
 	[Serializable]
 	sealed class EvalDelegateCreatorException : Exception { }
@@ -71,7 +71,7 @@ namespace dnSpy.Roslyn.Debugger.FilterExpressionEvaluator {
 			this.evalMethodName = evalMethodName;
 		}
 
-		static readonly Type[] evalDelegateParamTypes = new Type[] { typeof(string), typeof(ulong), typeof(string), typeof(ulong), typeof(string) };
+		static readonly Type[] evalDelegateParamTypes = new Type[] { typeof(string), typeof(int), typeof(string), typeof(ulong), typeof(string) };
 		public EvalDelegate CreateDelegate() {
 			var type = module.Find(evalClassName, isReflectionName: true);
 			Debug.Assert(type != null);
